backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
llama-index = "^0.9.19"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
pymongo = "^4.6.1"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
wikipedia = "^1.4.0"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"